+++ .formatter.exs
@@ -1,5 +1,8 @@
 # Used by "mix format"
 [
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
-  line_length: 120
+  line_length: 120,
+  locals_without_parens: [
+    assert_recent: :*
+  ]
 ]